+++ R/geom-quiver.r
@@ -11,8 +11,8 @@
 #' @examples
 #' library(ggplot2)
 #' # Quiver plots of mathematical functions
-#' expand.grid(x=seq(0,pi,pi/12), y=seq(0,pi,pi/12)) %>%
-#'   ggplot(aes(x=x,y=y,u=cos(x),v=sin(y))) +
+#' field <- expand.grid(x=seq(0,pi,pi/12), y=seq(0,pi,pi/12))
+#' ggplot(field, aes(x=x,y=y,u=cos(x),v=sin(y))) +
 #'   geom_quiver()
 #'
 #' # Removing automatic scaling
@@ -20,14 +20,13 @@
 #'   geom_quiver(vecsize=NULL) +
 #'   borders("state")
 #'
-#' \dontrun{
+#' @examplesIf requireNamespace("dplyr") && requireNamespace("ggmap")
 #' # Centering arrows is useful for plotting on maps.
 #' library(dplyr)
 #' library(ggmap)
 #' wind_data <- wind %>% filter(between(lon, -96, -93) & between(lat, 28.7, 30))
 #' qmplot(lon, lat, data=wind_data, extent="panel", geom = "blank", zoom=8, maptype = "toner-lite") +
 #'   geom_quiver(aes(u=delta_lon, v=delta_lat, colour = spd), center=TRUE)
-#' }
 #'
 #' @importFrom ggplot2 layer
 #'
@@ -65,18 +64,14 @@ geom_quiver <- function(mapping = NULL, data = NULL,
 GeomQuiver <- ggproto(
   "GeomQuiver", ggplot2::GeomSegment,
   draw_panel = function(data, panel_params, coord, arrow = NULL, lineend = "butt", na.rm = FALSE) {
-    trans <- CoordCartesian$transform(data, panel_params) %>%
-      mutate(arrowsize = sqrt((x - xend) ^ 2 + (y - yend) ^ 2) * 0.5)
-    grid::segmentsGrob(
-      trans$x, trans$y, trans$xend, trans$yend,
-      default.units = "native",
-      gp = grid::gpar(
-        col = alpha(trans$colour, trans$alpha),
-        lwd = trans$size * .pt,
-        lty = trans$linetype,
-        lineend = lineend
-      ),
-      arrow = grid::arrow(length = unit(trans$arrowsize, "npc"))
+    # Compute appropriate arrow size
+    data$arrowsize <- with(ggplot2::CoordCartesian$transform(data, panel_params),
+                           sqrt((x - xend) ^ 2 + (y - yend) ^ 2) * 0.5)
+    # Re-use segments to produce arrows
+    ggplot2::GeomSegment$draw_panel(
+      data, panel_params, coord,
+      arrow = grid::arrow(length = unit(data$arrowsize, "npc")),
+      lineend = lineend
     )
   }
 )