+++ NEWS.md
@@ -1,3 +1,8 @@
+# ggquiver 0.3.0 (29 October 2021)
+* Improved handling of non cartesian coordinate systems.
+* Fixed issue with plotting arrows on `ggmap::qmplot()` (#7).
+* Performance improvements.
+
 # Version 0.2.0 (5 February 2019)
 * Changed ggplot2 to import
 * Updated README to reflect ggmap changes